+++ src/index.ts
@@ -1,4 +1,5 @@
+/** Entrypoint the bundle */
+
 import * as Hotmap from './main';
 
 (window as any).Hotmap = Hotmap;
-